Login-Auth/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pwdlib[argon2]
python-dotenv
fastapi-mail
email-validator
